Created 'operating_system' folder with 'scheduling' and 'process_synchronization' subfolders
